update - categoria

+++ src/main/java/ZtechAplication/repository/CategoriaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import ZtechAplication.model.Categoria;
 import ZtechAplication.model.Cliente;
 
-public interface CategoriaRepository extends CrudRepository<Categoria, Long>{
+public interface CategoriaRepository extends CrudRepository<Categoria, Integer>{
 	Optional<Categoria> findByNome(String cate);
 
 }
